requirements.txt
disnake[voice]
pyyaml
pyrebase4
revChatGPT

